Backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
langchain
langchain-google-genai
google-generativeai
faiss-cpu
langchain-community
sentence-transformers
torch